Roomba: Check if Arduino has a HW serial or SW serial
Some Arduino models, such as the UNO, do not offer a hardware serial and simply share their RX(0) and TX(1) pins with their FTDI usb to serial chip. The default Serial class in these models is an instance of Serial_, not HardwareSerial. Some models, such as the Arduino Leonardo, do offer a seperate HW serial communication on pins RX(0) and TX(1) via the Serial1 object, which is an instance of HardwareSerial. In these models, the "Serial" object is explicitly reserved for the FTDI chip only! Finally, in some more powerful Arduino models, such as SAM cores, the default "Serial" object is an instance of HardwareSerial.

Signed-off-by: CTXz <ctx.xda@gmail.com>